tcg: fence merging optimization pass
If two fences in the TCG IR have no memory access between them, the can be
merged to a single fence enfocing what both fences enforced. In practice, we
change the type of the first fence to an OR of the types of both fences, and
delete the second fence. This pass happens after the liveness pass.